optimize && separate Home file
